remove builder add TintedProgressDialog that extends from AlertDialog
